requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
mysqlclient==2.2.4
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
